Fix bug in quality index sunburst
The resulting sector sizes have rounding errors due to the floating point operations.
But, the sunburst plot needs the total value for node (parent) bigger or equal than the sum of its children.
In some cases, the values were slightly smaller and the plot was not shown.
This is a fix that by adding a small margin to each node.